ScheduleViewController переделан под адекватный вид. Вынесенo лишнее из LC, переделан UI
