--- ScheduleViewController/ScheduleViewControllerO.swift
@@ -1,237 +1,0 @@
-//
-//  ScheduleViewController.swift
-//  Tracker
-//
-//  Created by Кирилл Марьясов on 17.07.2024.
-//
-
-import UIKit
-
-protocol ScheduleViewControllerDelegate: AnyObject {
-    func didSelectDays(_ days: [DayOfWeek])
-}
-
-protocol ScheduleCellDelegate: AnyObject {
-    func switchButtonClicked(to isSelected: Bool, of weekDay: DayOfWeek)
-}
-
-class ScheduleViewController: UIViewController {
-    
-    weak var delegate: ScheduleViewControllerDelegate?
-    private var selectedWeekDays: Set<DayOfWeek> = []
-
-    // MARK: - UI Elements
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.text = "Расписание"
-        label.font = UIFont.systemFont(ofSize: 19)
-        label.textAlignment = .center
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private let tableView: UITableView = {
-        let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.register(SwitchTableViewCell.self, forCellReuseIdentifier: "switchCell")
-        tableView.layer.cornerRadius = 10
-        tableView.separatorStyle = .singleLine
-        tableView.isScrollEnabled = false
-        tableView.tableHeaderView = nil
-        tableView.sectionHeaderHeight = 0
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-        return tableView
-    }()
-    
-    private let doneButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.setTitle("Готово", for: .normal)
-        button.setTitleColor(.ypWhite, for: .normal)
-        button.backgroundColor = .ypBlack
-        button.layer.cornerRadius = 10
-        button.translatesAutoresizingMaskIntoConstraints = false
-        return button
-    }()
-    
-    // MARK: - Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        tableView.dataSource = self
-        tableView.delegate = self
-        setUpView()
-        
-    }
-    
-    deinit {
-        selectedWeekDays.removeAll()
-        print("SelectedWeekDays array cleared!")
-    }
-
-    // MARK: - Setup UI
-    private func setUpView() {
-        view.backgroundColor = .ypWhite
-        
-        view.addSubview(titleLabel)
-        view.addSubview(doneButton)
-        view.addSubview(tableView)
-        
-        doneButton.addTarget(self, action: #selector(doneButtonTapped), for: .touchUpInside)
-        
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
-            titleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            titleLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
-            
-            doneButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
-            doneButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            doneButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            doneButton.heightAnchor.constraint(equalToConstant: 60),
-            doneButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: 16),
-            
-            tableView.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 30),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
-            tableView.heightAnchor.constraint(equalToConstant: 525)
-        ])
-    }
-    
-    @objc private func doneButtonTapped() {
-        let weekDays = Array(selectedWeekDays)
-        print("Selected days: \(weekDays)")
-        delegate?.didSelectDays(weekDays)
-        print("Delegate method called")
-        self.dismiss(animated: true) {
-            print("ScheduleViewController dismissed")
-        }
-    }
-}
-
-extension ScheduleViewController: ScheduleCellDelegate {
-    func switchButtonClicked(to isSelected: Bool, of weekDay: DayOfWeek) {
-        if isSelected {
-            selectedWeekDays.insert(weekDay)
-        } else {
-            selectedWeekDays.remove(weekDay)
-        }
-    }
-}
-
-extension ScheduleViewController: UITableViewDataSource {
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return DayOfWeek.allCases.count
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 75
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "switchCell", for: indexPath) as? SwitchTableViewCell else { fatalError("Не удалось найти ячейку ScheduleCell") }
-        cell.delegate = self
-        cell.selectionStyle = .none
-        let weekDay = DayOfWeek.allCases[indexPath.row]
-        cell.configureCell(
-            with: weekDay,
-            isLastCell: indexPath.row == 6,
-            isSelected: selectedWeekDays.contains(weekDay)
-        )
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 0.001
-    }
-    
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return UIView(frame: .zero)
-    }
-}
-
-extension ScheduleViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-}
-
-class SwitchTableViewCell: UITableViewCell {
-    weak var delegate: ScheduleCellDelegate?
-    
-    private var weekDay: DayOfWeek?
-    
-    private lazy var cellTitleLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = .ypBlack
-        label.font = UIFont.systemFont(ofSize: 17, weight: .regular)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private lazy var switchButton: UISwitch = {
-        let switcher = UISwitch()
-        switcher.onTintColor = .systemBlue
-        switcher.addTarget(self, action: #selector(switchButtonTapped(_:)), for: .valueChanged)
-        switcher.translatesAutoresizingMaskIntoConstraints = false
-        return switcher
-    }()
-    
-    private lazy var separatorView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .ypGray
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setupViews()
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func configureCell(with weekDay: DayOfWeek, isLastCell: Bool, isSelected: Bool) {
-        self.weekDay = weekDay
-        cellTitleLabel.text = weekDay.rawValue
-        separatorView.isHidden = isLastCell
-        switchButton.isOn = isSelected
-    }
-    
-    private func setupViews() {
-        contentView.backgroundColor = .ypBackground
-        contentView.addSubview(cellTitleLabel)
-        contentView.addSubview(switchButton)
-        contentView.addSubview(separatorView)
-    }
-    
-    private func setupConstraints() {
-        let contentViewHeightConstraint = contentView.heightAnchor.constraint(equalToConstant: 75)
-        contentViewHeightConstraint.isActive = true
-        
-        NSLayoutConstraint.activate([
-            cellTitleLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 26),
-            cellTitleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            cellTitleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            cellTitleLabel.heightAnchor.constraint(equalToConstant: 22),
-            
-            contentView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor),
-            contentView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor),
-            contentView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor),
-            
-            switchButton.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            switchButton.centerYAnchor.constraint(equalTo: cellTitleLabel.centerYAnchor),
-            
-            separatorView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            separatorView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            separatorView.heightAnchor.constraint(equalToConstant: 0.5),
-            separatorView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
-        ])
-    }
-
-    @objc private func switchButtonTapped(_ sender: UISwitch) {
-        guard let weekDay = weekDay else { return }
-        delegate?.switchButtonClicked(to: sender.isOn, of: weekDay)
-    }
-}
