+++ TrackerModels/Week.swift
@@ -1,0 +1,18 @@
+//
+//  Week.swift
+//  Tracker
+//
+//  Created by Кирилл Марьясов on 11.09.2024.
+//
+
+import Foundation
+
+enum Weekday: String {
+  case monday = "Пн"
+  case tuesday = "Вт"
+  case wednesday = "Ср"
+  case thursday = "Чт"
+  case friday = "Пт"
+  case saturday = "Cб"
+  case sunday = "Вск"
+}
